Fixed RPC ForwardPacket
packet is always padded to 4 bytes
